feat(PLAYING-MUSIC): actually support playing music!
definitely rough around the edges, but now we can build a queue, play the queue on the correct device, and still use the website while its playing. the queue can be added on to while playing and the website websocket messages for now playing and progress works still. the queue websocket thing died but thatll get fixed later